[feature] Present two method of linear regression
