+++ hw01/linear_regression.py
@@ -1,208 +1,132 @@
-#!python2.7
-import os
-from collections import OrderedDict
-from datetime import timedelta
+#!/usr/local/bin/python2.7
+import math
 
 import numpy as np
-import pandas as pd
-
-from gradient_descent import GradientDescent
-
-PATH_GIVEN_TRAIN = './given/train.csv'
-PATH_GIVEN_TEST = './given/test_X.csv'
-PATH_TRAIN_DATA = './train_data.csv'
-PATH_VALID_DATA = './valid_data.csv'
-PATH_RESULT = 'result/linear_regression.csv'
 
 
-def _create_tidy_training_set():
-    # read raw data and label it
-    # 1...24 mean hour of that day
-    colnames = ['Date', 'Site', 'Item'] + map(str, range(24))
-    df = pd.read_csv(PATH_GIVEN_TRAIN, names=colnames, skiprows=1)
+class LinearRegression(object):
+    def __init__(self):
+        self.__wt = None
+        self.__b = None
+        self.__X = None
+        self.__y = None
 
-    # remove column 'Site'
-    df = df.loc[:, ['Date', 'Item'] + map(str, range(24))]
+    @property
+    def wt(self):
+        return self.__wt
 
-    # snapshot:
-    # colums: Date, Item, 1...24(hour)
+    @property
+    def b(self):
+        return self.__b
 
-    # melt 'Hour' to column
-    df = pd.melt(df,
-                 id_vars=['Date', 'Item'],
-                 value_vars=map(str, range(24)),
-                 var_name='Hour',
-                 value_name='Value')
+    def train_by_pseudo_inverse(self, X, y, alpha=0, validate_data=None):  # noqa: N803
+        self._check_data(X, y)
 
-    # snapshot:
-    # colums: Date, Item, Hour, Value
+        self.__X = X
+        self.__y = y
 
-    # generate 'Datetime'
-    df['Datetime'] = pd.to_datetime(df.Date + ' ' + df.Hour + ':00:00')
-    df = df.loc[:, ['Datetime', 'Item', 'Value']]
+        num_data = X.shape[0]
+        dim_wt = X.shape[1]
 
-    # snapshot:
-    # colums: Datetime, Item, Value
+        argu_X = np.hstack((np.ones(num_data).reshape(num_data, 1), X))
 
-    # replace NR to 0
-    df.loc[df.Value == 'NR', 'Value'] = 0
+        # calculate pseudo-inverse
+        A_plus = np.dot(np.linalg.inv(np.dot(argu_X.T, argu_X)+alpha*np.eye(dim_wt+1)), argu_X.T)
 
-    # change 'Value' type
-    df['Value'] = df['Value'].astype(float)
+        wt_b = np.dot(A_plus, y)
+        self.__b = wt_b[0]
+        self.__wt = wt_b[1:]
 
-    # pivot 'Item' to columns
-    df = df.pivot_table(values='Value', index='Datetime', columns='Item', aggfunc='sum')
+        if validate_data:
+            print 'Pseudo-Inverse: err = {:.6f} validate = {:.6f}'.format(
+                self.err_insample(),
+                self.err(validate_data[0], validate_data[1]))
+        else:
+            print 'Pseudo-Inverse: err = {:.6f}'.format(self.err_insample())
 
-    # snapshot:
-    # index: Datetime
-    # colums: AMB_TEMP, CH4, CO,..., PM2.5,..., WS_HR
+    def train_by_gradient_descent(self, X, y, init_wt=np.array([]), init_b=0,  # noqa: N803
+              rate=0.01, alpha=0, epoch=1000, batch=None, validate_data=None):
+        self._check_data(X, y)
 
-    return df
+        if init_wt.size == 0:
+            init_wt = np.zeros(X.shape[1])
 
+        self.__X = X
+        self.__y = y
 
-def _split_df_into_train_and_valid(df):
-    df_valid = df.loc[df.index.month == 12, :]
-    df_train = df.loc[df.index.month != 12, :]
-    return (df_train, df_valid)
+        self.__wt = init_wt
+        self.__b = init_b
 
+        num_data = X.shape[0]
 
-def _convert_to_problem_regression_form(df):
-    """Convert to the problem regression form.
+        if not batch:
+            batch = num_data
 
-    The problem is that using parameters among continue 9 hours to predict PM2.5 on next hour.
+        tot_batch = int(math.ceil(float(num_data) / float(batch)))
+        if validate_data:
+            for i in range(epoch):
+                for j in range(tot_batch):
+                    batch_X = self.__X[j*batch:min(num_data, (j+1)*batch), :]
+                    batch_y = self.__y[j*batch:min(num_data, (j+1)*batch)]
+                    self.__b, self.__wt = self._gd_update(
+                        batch_X, batch_y, self.__wt, self.__b, rate, alpha)
+                print 'Epoch {:5d}: err = {:.6f} validate = {:.6f}'.format(
+                    i+1,
+                    self.err_insample(),
+                    self.err(validate_data[0], validate_data[1]))
+        else:
+            for i in range(epoch):
+                for j in range(tot_batch):
+                    batch_X = self.__X[j*batch:min(num_data, (j+1)*batch), :]
+                    batch_y = self.__y[j*batch:min(num_data, (j+1)*batch)]
+                    self.__b, self.__wt = self._gd_update(
+                        batch_X, batch_y, self.__wt, self.__b, rate, alpha)
+                print 'Epoch {:5d}: err = {:.6f}'.format(i+1, self.err_insample())
 
-    Args:
-        df (pd.Dataframe): `index` is datetime
-                           `columns` include parameters AMB_TEMP, CH4, CO,..., PM2.5,..., WS_HR
+    def _gd_update(self, X, y, wt, b, rate, alpha):  # noqa: N803
+        num_data = X.shape[0]
 
-    """
-    data = OrderedDict()
+        # y_pred
+        y_pred = np.sum(wt * X, axis=1) + b
+        # y_pred - y
+        y_diff = y_pred - y
 
-    # create columns of `data`
-    param_list = df.columns.tolist()
-    for i in range(9):
-        for param in param_list:
-            data['{:02d}h__{}'.format(i+1, param)] = []
-    data['10h__PM2.5'] = []
+        # b update
+        b_gradient = np.sum(y_diff)/num_data
+        new_b = b - rate * b_gradient
+        # wt update
+        new_wt = []
+        for i, w in enumerate(wt):
+            w_gradient = np.sum(X[:, i] * y_diff)/num_data + alpha * w
+            new_wt.append(w - rate * w_gradient)
+        new_wt = np.array(new_wt, dtype='float64')
+        return (new_b, new_wt)
 
-    # add content into `data`
-    datetime_list = df.index
-    d1h = timedelta(hours=1)
-    for m in pd.unique(datetime_list.month):
-        for timestamp in (df.loc[df.index.month == m, :]).index:
-            start, end = timestamp, timestamp + 9*d1h
-            sub_df = df.loc[(start <= df.index) & (df.index <= end), :]
-            if sub_df.shape[0] == 10:
-                for i in range(9):
-                    for param in param_list:
-                        data['{:02d}h__{}'.format(i+1, param)].append(
-                            sub_df.loc[timestamp+i*d1h, param])
-                data['10h__PM2.5'].append(sub_df.loc[timestamp + 9 * d1h, 'PM2.5'])
+    def predict(self, X):  # noqa: N803
+        if type(X) == list:
+            X = np.array(X, dtype='float64')
 
-    return pd.DataFrame(data)
+        if X.shape[1] != self.__wt.shape[0]:
+            raise ValueError('shape of input x does not match shape of weight')
 
+        return np.sum(self.__wt * X, axis=1)+self.__b
 
-def preprocess_training_set():
-    if os.path.isfile(PATH_TRAIN_DATA) and os.path.isfile(PATH_VALID_DATA):
-        train_df = pd.read_csv(PATH_TRAIN_DATA)
-        valid_df = pd.read_csv(PATH_VALID_DATA)
-    else:
-        df = _create_tidy_training_set()
-        tmp_train_df, tmp_valid_df = _split_df_into_train_and_valid(df)
+    def err_insample(self):
+        if self.__X.size == 0 or self.__y.size == 0:
+            raise RuntimeError('in-sample data not found')
 
-        train_df = _convert_to_problem_regression_form(tmp_train_df)
-        train_df.to_csv(PATH_TRAIN_DATA, index=None)
-        valid_df = _convert_to_problem_regression_form(tmp_valid_df)
-        valid_df.to_csv(PATH_VALID_DATA, index=None)
+        return self.err(self.__X, self.__y)
 
-    train_X = np.array(train_df.loc[:, train_df.columns != '10h__PM2.5'])
-    train_y = np.array(train_df.loc[:, '10h__PM2.5'])
+    def err(self, X, y):  # noqa: N803
+        self._check_data(X, y)
 
-    valid_X = np.array(valid_df.loc[:, valid_df.columns != '10h__PM2.5'])
-    valid_y = np.array(valid_df.loc[:, '10h__PM2.5'])
+        if self.__wt.size == 0 or self.__b.size == 0:
+            raise RuntimeError('you should train model first')
+        y_pred = np.sum(self.__wt * X, axis=1) + self.__b
 
-    col_name = (train_df.loc[:, train_df.columns != '10h__PM2.5']).columns
+        err = (np.sum((y - y_pred)**2) / y.shape[0])**0.5
+        return round(err, 8)
 
-    return (train_X, train_y, valid_X, valid_y, col_name)
-
-
-def preprocess_testing_set(col_name):
-    col_names = ['ID', 'Item'] + map(lambda x: '{:02d}h'.format(x), range(1, 10))
-    test_df = pd.read_csv(PATH_GIVEN_TEST, names=col_names, header=None)
-
-    # snapshot:
-    # columns: ID, Item, 01h, 02h, ..., 09h
-
-    # replace NR to 0
-    for col in map(lambda x: '{:02d}h'.format(x), range(1, 10)):
-        test_df.loc[(test_df.Item == 'RAINFALL') & (test_df[col] == 'NR'), col] = 0
-
-    # ['ID','Item','Hour','Value'] form
-    test_df = test_df.pivot_table(index=['ID', 'Item'], aggfunc='sum')
-    test_df = test_df.stack()
-    test_df = test_df.reset_index()
-    test_df.columns = ['ID', 'Item', 'Hour', 'Value']
-
-    # snapshot:
-    # columns: ID, Item, Hour, Value
-
-    # combine 'Hour' and 'Item' to 'Col'
-    test_df['Col'] = test_df.Hour + '__' + test_df.Item
-    test_df = test_df[['ID', 'Col', 'Value']]
-
-    # snapshot:
-    # columns: ID, Col, Value
-
-    # pivot 'Col' to columns
-    test_df = test_df.pivot_table(values='Value', index='ID', columns='Col', aggfunc='sum')
-
-    # snapshot:
-    # index: ID
-    # columns: 01h__AMB_TEMP, 01h__CH4, 01h__CO,..., 09h__WS_HR
-
-    # re-order
-    test_df['ID_Num'] = test_df.index.str.replace('id_', '').astype('int')
-    test_df = test_df.sort_values(by='ID_Num')
-    test_df = test_df.drop('ID_Num', axis=1)
-
-    # snapshot:
-    # index: ID
-    # columns: ID_Num, 01h__AMB_TEMP, 01h__CH4, 01h__CO,..., 09h__WS_HR
-
-    test_X = np.array(test_df[col_name], dtype='float64')
-    ids = np.array(test_df.index, dtype='str')
-    return (test_X, ids)
-
-
-def main():
-    train_X, train_y, valid_X, valid_y, col_name = preprocess_training_set()
-
-    # ### gradient descent
-    gd = GradientDescent()
-
-    # method: pseudo_inverse
-    gd.train_by_pseudo_inverse(train_X, train_y, alpha=0.5, validate_data=(valid_X, valid_y))
-
-    # method: gradient desecent
-    # init_wt = gd.wt
-    # init_b  = gd.b
-    #
-    # gd.train(train_X,train_y,epoch=10,rate=0.000001,batch=100,alpha=0.00000001,
-    #    init_wt=np.array(init_wt),init_b=init_b,
-    #    validate_data = (valid_X,valid_y))
-
-    # ### predict
-    test_X, ids = preprocess_testing_set(col_name)
-    pred_y = gd.predict(test_X)
-
-    result = list()
-    for i in range(ids.shape[0]):
-        result.append([ids[i], pred_y[i]])
-
-    # output
-    with open(PATH_RESULT, 'w') as fw:
-        for id, pred in result:
-            fw.write('{id},{pred}\n'.format(id=id, pred=pred))
-
-if __name__ == '__main__':
-    main()
+    def _check_data(self, X, y):  # noqa: N803
+        if X.shape[0] != y.shape[0]:
+            raise ValueError('shape of X and y do not match')
