+++ hw01/result/gradient_descent.csv
@@ -1,0 +1,240 @@
+id_0,24.6561773638
+id_1,63.3106818979
+id_2,18.7539095531
+id_3,26.9030837746
+id_4,8.91199362475
+id_5,28.4090666088
+id_6,39.1211805422
+id_7,15.5936316413
+id_8,49.9013801176
+id_9,31.5281713037
+id_10,36.5654916793
+id_11,55.2079998239
+id_12,34.3621992038
+id_13,34.8368922674
+id_14,28.4545829735
+id_15,9.90480672962
+id_16,50.3381034481
+id_17,21.6328447175
+id_18,14.2736255109
+id_19,16.6300896279
+id_20,14.9722301703
+id_21,28.3320235473
+id_22,37.4502781661
+id_23,9.58603791768
+id_24,59.4994564266
+id_25,44.3384126029
+id_26,24.236477433
+id_27,6.86905896192
+id_28,27.8436585809
+id_29,34.6897060144
+id_30,15.9877132738
+id_31,10.48882965
+id_32,14.8146190913
+id_33,6.51684193348
+id_34,14.8336665137
+id_35,42.015155535
+id_36,21.4802902189
+id_37,25.8011935022
+id_38,10.4133681636
+id_39,55.5976766218
+id_40,43.2603464845
+id_41,19.8127652707
+id_42,51.2913264787
+id_43,17.4866226229
+id_44,41.3719903277
+id_45,48.4185058093
+id_46,56.9007913351
+id_47,39.8449182698
+id_48,52.4849376571
+id_49,36.1093461215
+id_50,30.0957948458
+id_51,20.9517879742
+id_52,42.0346368161
+id_53,24.3853866833
+id_54,33.480677887
+id_55,37.3311471569
+id_56,22.880712236
+id_57,19.7287848747
+id_58,35.1817056317
+id_59,31.341439519
+id_60,10.1911400368
+id_61,23.8290913323
+id_62,29.473573431
+id_63,47.7392135628
+id_64,48.0407157342
+id_65,24.9060487649
+id_66,35.5712092252
+id_67,50.1436905828
+id_68,27.2042937387
+id_69,37.2042415863
+id_70,62.7328672953
+id_71,54.5024553777
+id_72,40.0008158384
+id_73,13.3827671157
+id_74,11.885812917
+id_75,33.9105625233
+id_76,61.4398075438
+id_77,1.72732055926
+id_78,41.3067789764
+id_79,40.6911018537
+id_80,8.0928501756
+id_81,19.761526509
+id_82,4.89783019143
+id_83,30.6207069534
+id_84,6.27116441793
+id_85,4.15515320498
+id_86,16.1826892471
+id_87,10.8497634589
+id_88,15.8336619092
+id_89,20.6394663184
+id_90,34.6899272842
+id_91,11.1418830802
+id_92,10.0642328535
+id_93,10.629169051
+id_94,34.362724878
+id_95,13.4825091606
+id_96,15.2942287876
+id_97,8.21822282838
+id_98,9.21802093484
+id_99,5.13721396059
+id_100,8.91172591218
+id_101,17.644160266
+id_102,9.78934931243
+id_103,8.85875459041
+id_104,12.5765035832
+id_105,34.7418456208
+id_106,20.8348490438
+id_107,7.20107222111
+id_108,17.3609484441
+id_109,15.9886209601
+id_110,9.80123117812
+id_111,12.9937385709
+id_112,14.9039267655
+id_113,22.3291238988
+id_114,19.7533657201
+id_115,14.5045505955
+id_116,8.8533035212
+id_117,40.2497836493
+id_118,22.6647976896
+id_119,0.592839050749
+id_120,0.580604525422
+id_121,9.35279977531
+id_122,11.6965144566
+id_123,10.7018137412
+id_124,6.40588445898
+id_125,24.1348335973
+id_126,23.0704921741
+id_127,8.28014125509
+id_128,17.9590841433
+id_129,12.9444215458
+id_130,1.04822651174
+id_131,30.2881028149
+id_132,15.2096103049
+id_133,22.6876247003
+id_134,26.0557207153
+id_135,24.4309014317
+id_136,9.10829317778
+id_137,22.6144268058
+id_138,76.8233210087
+id_139,63.9979641903
+id_140,27.814380352
+id_141,18.2946730957
+id_142,22.9725260686
+id_143,6.12790412388
+id_144,13.719033597
+id_145,9.13910773733
+id_146,17.8215843336
+id_147,6.9127959187
+id_148,9.5116970247
+id_149,14.2474461444
+id_150,42.8811441903
+id_151,15.1936742453
+id_152,20.2080996468
+id_153,12.7052613283
+id_154,13.3568937369
+id_155,22.5787598094
+id_156,21.0125607096
+id_157,35.8478004176
+id_158,3.09824984557
+id_159,20.4392128202
+id_160,39.7711724949
+id_161,17.9569210887
+id_162,57.7223398532
+id_163,15.3521358334
+id_164,32.9300466276
+id_165,24.5405318175
+id_166,21.3790226602
+id_167,0.201557520894
+id_168,66.4830333077
+id_169,14.7553263388
+id_170,14.8771681454
+id_171,68.273738626
+id_172,18.9149174907
+id_173,2.71209401405
+id_174,22.6714903191
+id_175,-0.0838744754955
+id_176,7.95021315089
+id_177,16.1068443278
+id_178,21.2151713115
+id_179,20.7448747793
+id_180,40.9009338397
+id_181,34.4909606287
+id_182,30.6153064351
+id_183,14.063071991
+id_184,6.4858349883
+id_185,40.6668044751
+id_186,33.6191661037
+id_187,33.0632861607
+id_188,12.9548315231
+id_189,23.5810346769
+id_190,13.664475865
+id_191,14.7887147563
+id_192,27.9513455341
+id_193,74.1401578744
+id_194,31.8919761296
+id_195,9.72149462607
+id_196,59.3220028946
+id_197,18.0480118202
+id_198,18.4634604127
+id_199,19.7527841563
+id_200,119.128128124
+id_201,21.8107165688
+id_202,57.2169288313
+id_203,3.70184573466
+id_204,56.1174588889
+id_205,31.8503228658
+id_206,13.3178924227
+id_207,15.2276849862
+id_208,58.3891080078
+id_209,13.6753620901
+id_210,39.2638675969
+id_211,65.6759833053
+id_212,19.2788212711
+id_213,112.033023748
+id_214,39.8489372268
+id_215,72.9016537224
+id_216,87.3236335279
+id_217,68.5225061235
+id_218,31.7944970531
+id_219,22.6912948937
+id_220,29.5793080753
+id_221,54.0350575851
+id_222,24.7721805652
+id_223,36.6817455764
+id_224,43.1854872411
+id_225,9.01428616394
+id_226,5.31553246655
+id_227,80.236264761
+id_228,27.4155443652
+id_229,1.96561071023
+id_230,38.9916743656
+id_231,12.9939076934
+id_232,21.4587605919
+id_233,6.01081100287
+id_234,14.6998529197
+id_235,22.6985035884
+id_236,14.2815910042
+id_237,4.37020982284
+id_238,23.3211892036
+id_239,12.1198561242
