--- hw01/result/linear_regression.csv
@@ -1,240 +1,0 @@
-id_0,25.2556619606
-id_1,63.6505983675
-id_2,18.9728902917
-id_3,29.390804962
-id_4,9.56471181797
-id_5,33.2409863562
-id_6,40.8996415339
-id_7,18.2230019353
-id_8,53.2994099278
-id_9,32.5177831941
-id_10,38.8155322397
-id_11,59.157687411
-id_12,35.8703355586
-id_13,41.7291418511
-id_14,26.6079415546
-id_15,13.0020085786
-id_16,52.1664560166
-id_17,24.4893425043
-id_18,18.6758697474
-id_19,20.1410490894
-id_20,16.053266956
-id_21,32.0708037903
-id_22,39.5492528659
-id_23,11.5359844942
-id_24,61.25908573
-id_25,46.0641530842
-id_26,28.8057303163
-id_27,6.06525526605
-id_28,28.0540497705
-id_29,37.8889386389
-id_30,18.2944879485
-id_31,11.7166594128
-id_32,16.6862222536
-id_33,7.4178905864
-id_34,15.039569429
-id_35,41.9515203002
-id_36,22.0112760361
-id_37,28.0903224031
-id_38,4.3930643095
-id_39,59.8029944912
-id_40,47.9881527043
-id_41,20.678974618
-id_42,52.5157943652
-id_43,20.0034327755
-id_44,44.8447713765
-id_45,48.7032114661
-id_46,57.6241606757
-id_47,41.7237000637
-id_48,55.9261218003
-id_49,39.2407435636
-id_50,31.2669732578
-id_51,24.1280077352
-id_52,39.7093342877
-id_53,23.5779430578
-id_54,37.136178765
-id_55,38.81108313
-id_56,25.9523029194
-id_57,22.1323677994
-id_58,35.821861925
-id_59,31.8446686067
-id_60,13.1350235553
-id_61,21.7776339604
-id_62,32.5120208962
-id_63,53.1362390875
-id_64,51.3431560078
-id_65,25.824054577
-id_66,38.0108696379
-id_67,53.3486030341
-id_68,28.7997334197
-id_69,39.9668673874
-id_70,64.3472446205
-id_71,55.6050052785
-id_72,41.2192435713
-id_73,14.115387161
-id_74,13.4716357657
-id_75,35.9306252831
-id_76,62.4544345124
-id_77,2.88853580084
-id_78,40.7784654824
-id_79,42.6693354126
-id_80,7.89544836334
-id_81,22.878404885
-id_82,4.88030897371
-id_83,30.5642722321
-id_84,9.50068789887
-id_85,5.5448520801
-id_86,17.1890772922
-id_87,10.4251443507
-id_88,20.7884310171
-id_89,25.2296461615
-id_90,35.1742007056
-id_91,11.7778863895
-id_92,12.8069170202
-id_93,9.63584382269
-id_94,37.3699814335
-id_95,15.1348408657
-id_96,16.3010465835
-id_97,8.88613981189
-id_98,11.9737208245
-id_99,5.08645462806
-id_100,9.18549065516
-id_101,17.9377597824
-id_102,12.5123438975
-id_103,8.0780976977
-id_104,12.9367687432
-id_105,35.9922241846
-id_106,26.0059234111
-id_107,7.53945207058
-id_108,17.051612911
-id_109,15.2403753387
-id_110,12.1698031525
-id_111,17.3756528529
-id_112,16.0907593395
-id_113,26.6444788031
-id_114,19.1637779135
-id_115,16.6883121265
-id_116,6.02787579188
-id_117,45.8325285238
-id_118,26.5756678169
-id_119,0.557818389293
-id_120,-1.06665452105
-id_121,9.16911154788
-id_122,12.9193878647
-id_123,13.3141575869
-id_124,7.63226381955
-id_125,24.9426582639
-id_126,20.5470894328
-id_127,10.3830546523
-id_128,20.7856737708
-id_129,13.6778744197
-id_130,-1.31590796634
-id_131,32.7622731972
-id_132,13.8505340326
-id_133,20.5531764564
-id_134,26.2884253452
-id_135,23.8943474301
-id_136,9.92619435724
-id_137,21.2643265844
-id_138,78.4073837662
-id_139,60.4849548217
-id_140,26.8241621832
-id_141,18.3198936146
-id_142,24.4673127346
-id_143,7.27765350633
-id_144,11.9080316001
-id_145,10.3840886715
-id_146,19.368124791
-id_147,8.68959631993
-id_148,13.1336350589
-id_149,15.0831917497
-id_150,46.0607469251
-id_151,21.7000419435
-id_152,21.7387762643
-id_153,15.1069387452
-id_154,14.0403639756
-id_155,25.6379146565
-id_156,25.5481072002
-id_157,36.1549821802
-id_158,5.20082342048
-id_159,23.3178417855
-id_160,42.1138199223
-id_161,19.7013684334
-id_162,64.5870812089
-id_163,17.0704853614
-id_164,33.6867275779
-id_165,23.4484342733
-id_166,25.1499927913
-id_167,-0.0209533727776
-id_168,66.2644151215
-id_169,15.1040044137
-id_170,16.2508823154
-id_171,68.0867607994
-id_172,20.0931510163
-id_173,2.03383908599
-id_174,23.7692970348
-id_175,-0.600592898824
-id_176,7.3297506494
-id_177,18.3787811756
-id_178,19.5413983298
-id_179,24.8625002206
-id_180,43.3440664092
-id_181,32.194816537
-id_182,27.7219144411
-id_183,12.4967082794
-id_184,6.40606012395
-id_185,42.0598467874
-id_186,35.9240348376
-id_187,34.762705432
-id_188,17.5154419345
-id_189,23.3277684564
-id_190,14.6349294163
-id_191,16.8805141238
-id_192,30.9986931544
-id_193,60.1398216619
-id_194,29.191015067
-id_195,10.5284033696
-id_196,63.5602984215
-id_197,16.1677945656
-id_198,20.2504848429
-id_199,21.0164752369
-id_200,123.216976391
-id_201,23.4690704269
-id_202,58.5238661767
-id_203,6.39587631712
-id_204,60.1200803361
-id_205,36.8864343584
-id_206,15.2621948517
-id_207,14.4329655455
-id_208,62.6353949312
-id_209,14.8641929967
-id_210,36.1475433608
-id_211,68.5861935974
-id_212,18.8182481559
-id_213,113.618221886
-id_214,41.4618905613
-id_215,78.7439945631
-id_216,91.2193408382
-id_217,72.2388517982
-id_218,31.0862789732
-id_219,19.1491934378
-id_220,29.806855022
-id_221,57.9854642664
-id_222,28.6212239548
-id_223,38.6345681094
-id_224,43.8606913838
-id_225,10.7390511384
-id_226,3.22833830268
-id_227,81.2768152926
-id_228,31.1542560546
-id_229,0.711966903439
-id_230,41.4423232022
-id_231,16.5369976311
-id_232,22.6638112027
-id_233,7.92489780189
-id_234,14.3519482224
-id_235,25.2086618792
-id_236,14.0313502954
-id_237,3.22754771688
-id_238,22.5076837578
-id_239,11.3388268397
