+++ hw01/result/pseudo_inverse.csv
@@ -1,0 +1,240 @@
+id_0,25.255661961
+id_1,63.6505983676
+id_2,18.9728902913
+id_3,29.3908049619
+id_4,9.56471181799
+id_5,33.2409863561
+id_6,40.8996415338
+id_7,18.2230019356
+id_8,53.2994099279
+id_9,32.5177831944
+id_10,38.8155322395
+id_11,59.1576874108
+id_12,35.8703355589
+id_13,41.7291418514
+id_14,26.6079415549
+id_15,13.0020085786
+id_16,52.1664560171
+id_17,24.4893425044
+id_18,18.6758697475
+id_19,20.1410490897
+id_20,16.0532669561
+id_21,32.0708037904
+id_22,39.5492528659
+id_23,11.5359844947
+id_24,61.2590857303
+id_25,46.0641530842
+id_26,28.8057303165
+id_27,6.06525526581
+id_28,28.0540497703
+id_29,37.8889386391
+id_30,18.2944879485
+id_31,11.7166594126
+id_32,16.6862222537
+id_33,7.41789058598
+id_34,15.0395694289
+id_35,41.9515202993
+id_36,22.0112760361
+id_37,28.0903224031
+id_38,4.39306430956
+id_39,59.8029944916
+id_40,47.9881527045
+id_41,20.6789746181
+id_42,52.5157943652
+id_43,20.0034327755
+id_44,44.8447713769
+id_45,48.7032114663
+id_46,57.6241606758
+id_47,41.7237000635
+id_48,55.9261218002
+id_49,39.2407435638
+id_50,31.2669732577
+id_51,24.1280077353
+id_52,39.7093342875
+id_53,23.5779430577
+id_54,37.136178765
+id_55,38.8110831302
+id_56,25.9523029194
+id_57,22.1323677996
+id_58,35.8218619252
+id_59,31.8446686071
+id_60,13.1350235549
+id_61,21.7776339599
+id_62,32.512020896
+id_63,53.1362390872
+id_64,51.3431560075
+id_65,25.8240545768
+id_66,38.0108696381
+id_67,53.3486030345
+id_68,28.7997334195
+id_69,39.9668673876
+id_70,64.3472446204
+id_71,55.6050052782
+id_72,41.2192435714
+id_73,14.1153871607
+id_74,13.4716357657
+id_75,35.9306252832
+id_76,62.4544345125
+id_77,2.88853580072
+id_78,40.7784654819
+id_79,42.6693354128
+id_80,7.89544836337
+id_81,22.8784048849
+id_82,4.88030897439
+id_83,30.5642722319
+id_84,9.50068789924
+id_85,5.54485208049
+id_86,17.1890772921
+id_87,10.4251443511
+id_88,20.788431017
+id_89,25.2296461615
+id_90,35.1742007056
+id_91,11.7778863897
+id_92,12.8069170203
+id_93,9.63584382248
+id_94,37.3699814331
+id_95,15.1348408655
+id_96,16.3010465836
+id_97,8.88613981177
+id_98,11.9737208249
+id_99,5.08645462813
+id_100,9.18549065542
+id_101,17.9377597827
+id_102,12.5123438977
+id_103,8.07809769747
+id_104,12.9367687434
+id_105,35.9922241844
+id_106,26.0059234112
+id_107,7.53945207088
+id_108,17.0516129111
+id_109,15.2403753386
+id_110,12.1698031524
+id_111,17.3756528529
+id_112,16.0907593395
+id_113,26.6444788033
+id_114,19.1637779137
+id_115,16.6883121265
+id_116,6.02787579167
+id_117,45.8325285237
+id_118,26.575667817
+id_119,0.55781838993
+id_120,-1.06665452077
+id_121,9.16911154775
+id_122,12.919387865
+id_123,13.314157587
+id_124,7.63226381955
+id_125,24.9426582637
+id_126,20.5470894328
+id_127,10.3830546525
+id_128,20.7856737707
+id_129,13.6778744199
+id_130,-1.3159079658
+id_131,32.7622731969
+id_132,13.8505340328
+id_133,20.553176456
+id_134,26.2884253452
+id_135,23.8943474301
+id_136,9.92619435697
+id_137,21.2643265846
+id_138,78.4073837659
+id_139,60.4849548215
+id_140,26.824162183
+id_141,18.3198936145
+id_142,24.4673127346
+id_143,7.2776535062
+id_144,11.9080315999
+id_145,10.3840886715
+id_146,19.3681247913
+id_147,8.68959631967
+id_148,13.1336350592
+id_149,15.0831917501
+id_150,46.0607469254
+id_151,21.7000419432
+id_152,21.7387762642
+id_153,15.1069387455
+id_154,14.0403639756
+id_155,25.6379146564
+id_156,25.5481072001
+id_157,36.15498218
+id_158,5.20082342075
+id_159,23.3178417856
+id_160,42.1138199224
+id_161,19.7013684334
+id_162,64.5870812091
+id_163,17.0704853614
+id_164,33.6867275776
+id_165,23.4484342733
+id_166,25.1499927916
+id_167,-0.0209533729331
+id_168,66.2644151213
+id_169,15.1040044139
+id_170,16.2508823152
+id_171,68.0867607994
+id_172,20.0931510162
+id_173,2.03383908595
+id_174,23.7692970341
+id_175,-0.600592898651
+id_176,7.32975064943
+id_177,18.3787811773
+id_178,19.5413983297
+id_179,24.8625002204
+id_180,43.3440664091
+id_181,32.194816537
+id_182,27.7219144414
+id_183,12.4967082791
+id_184,6.40606012388
+id_185,42.0598467875
+id_186,35.9240348375
+id_187,34.7627054321
+id_188,17.5154419346
+id_189,23.3277684563
+id_190,14.6349294164
+id_191,16.8805141235
+id_192,30.9986931545
+id_193,60.1398216595
+id_194,29.191015067
+id_195,10.5284033691
+id_196,63.5602984215
+id_197,16.1677945657
+id_198,20.2504848426
+id_199,21.016475237
+id_200,123.216976391
+id_201,23.4690704271
+id_202,58.5238661769
+id_203,6.39587631744
+id_204,60.1200803364
+id_205,36.8864343584
+id_206,15.2621948518
+id_207,14.4329655456
+id_208,62.6353949314
+id_209,14.8641929966
+id_210,36.1475433608
+id_211,68.5861935977
+id_212,18.8182481558
+id_213,113.618221886
+id_214,41.4618905614
+id_215,78.7439945629
+id_216,91.2193408385
+id_217,72.2388517987
+id_218,31.0862789734
+id_219,19.149193438
+id_220,29.8068550219
+id_221,57.9854642666
+id_222,28.6212239553
+id_223,38.6345681096
+id_224,43.8606913837
+id_225,10.7390511389
+id_226,3.22833830258
+id_227,81.2768152927
+id_228,31.154256055
+id_229,0.711966903788
+id_230,41.4423232026
+id_231,16.5369976309
+id_232,22.6638112026
+id_233,7.92489780185
+id_234,14.3519482218
+id_235,25.2086618789
+id_236,14.0313502956
+id_237,3.22754771708
+id_238,22.5076837576
+id_239,11.3388268396
